requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.30
psycopg==3.2.0
psycopg-binary==3.2.13
alembic==1.12.1
pydantic==2.10.0
python-dotenv==1.0.0
PyJWT==2.8.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
black
python-jose[cryptography]
kafka-python==2.3.0
